pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [
    "hatchling",
]

[project]
name = "aa-timezones"
description = "Time Zones Overview for Alliance Auth"
readme = "README.md"
keywords = [
    "allianceauth",
    "eveonline",
    "time",
    "timezones",
]
license = {file = "LICENSE"}
authors = [
    { name = "Peter Pfeufer", email = "develop@ppfeufer.de" },
]
requires-python = ">=3.8"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dynamic = [
    "version",
]
dependencies = [
    "allianceauth>=4.0.0a4",
    "allianceauth-app-utils>=1.25",
    "pytz>=2023.3.post1",
]
[project.optional-dependencies]
tests-allianceauth-latest = [
    "allianceauth-discordbot",
    "coverage",
    "django-webtest",
]
[project.urls]
Changelog = "https://github.com/ppfeufer/aa-timezones/blob/master/CHANGELOG.md"
Documentation = "https://github.com/ppfeufer/aa-timezones/blob/master/README.md"
Donations = "https://ko-fi.com/ppfeufer"
Homepage = "https://github.com/ppfeufer/aa-timezones"
Source = "https://github.com/ppfeufer/aa-timezones.git"
Tracker = "https://github.com/ppfeufer/aa-timezones/issues"

[tool.hatch.build]
include = [
    "/timezones",
]

[tool.hatch.version]
path = "timezones/__init__.py"
